ai/requirements.txt
# /ai/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
openai==0.28.1
httpx==0.25.0
python-multipart==0.0.6